==> Instacart Basket Analysis/02 Scripts/1. Data Import and Descriptive Analysis.ipynb <==
# IC DATA IMPORT AND DESCRIPTIVE ANALYSIS

## 1. Importing Libraries

## 2. Importing Data and Printing First/Last Rows

## 3. Printing Columns and Number of Rows/Columns

## 4. Statistical values for all the columns

## 5. Data types for all the columns

# -----------------------------------------------

## 1. Importing Libraries

# Import libraries:

import pandas as pd
import numpy as np
import os

## 2. Importing Data and Printing First/Last Rows

# Turning folder path into a string:

path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

# Importing orders.csv without 'eval_set':

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

df = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), usecols = vars_list)

# Printing first 5 rows of orders.csv:

df.head()

# Importing products.csv:

df_products = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'))

# Printing first 20 rows of orders.csv:

df_products.head(20)

# Printing last 35 rows of orders.csv:

df_products.tail(35)

## 3. Printing Columns and Number of Rows/Columns

# Printing columns names in products.csv

df_products.columns

# Printing number of rows and columns in products.csv

df_products.info()

## 4. Statistical values for all the columns

df_products.describe()

#### - The maximum value of the “aisle_id” column is: 134.000000

## 5. Data types for all the columns

df_products.dtypes

####  - The data type of the “department_id” column is: int64

df.shape

df_products.shape
