==> Instacart Basket Analysis/02 Scripts/2. Data Wrangling & Subsetting.ipynb <==
# DATA WRANGLING & SUBSETTING

# 1. Importing Libraries and Dataframes ('orders.csv' and 'products.csv')

# 2. Data Wrangling (Exercise)

## 2.1. Removing Columns

## 2.2. Finding Missing Values

## 2.3. Renaming Columns

## 2.4. Changing Data Type

## 2.5. Importing 'departments.csv'

## 2.6. Transposing

# 3. Subsetting

## 3.1. Creating Data Dictionary

## 3.2. Creating Subset

# 4. Data Wrangling & Subsetting (Task)

## 4.1. Changing column data type and name

## 4.2. Finding busiest hour and Determining meaning behind a value

## 4.3. Creating subset and finding all observations

## 4.4. Counting rows and and viewing all info + basic stats

# 5. Exporting 'orders_wrangled' and 'departments_wrangled'

# ---------------------------------------------------

# 1. Importing Libraries and Dataframes ('orders.csv' and 'products.csv')

# Importing the pandas, NumPy, and os libraries:

import pandas as pd
import numpy as np
import os

# Defining path:

path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

# Importing orders.csv:

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'))

# Importing products.csv:

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'))

# 2. Data Wrangling (Exercise)

## 2.1. Removing Columns

# Removing column 'eval_set' from orders.csv:

df_ords.drop(columns = ['eval_set'])

## 2.2. Finding Missing Values

# Finding missing value in 'days_sice_prior_order':

df_ords['days_since_prior_order'].value_counts(dropna = False)

## 2.3. Renaming Columns

# Renaming 'order_dow' column into 'orders_day_of_week':

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

## 2.4. Changing Data Type

# Changing data type of 'order_id' column in orders.csv:

df_ords['order_id'] = df_ords['order_id'].astype('str')

type('order_id')

df_ords['order_id'].dtype

## 2.5. Importing 'departments.csv'

# Importing 'departments.csv':

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'departments.csv'), index_col = False)

df_dep.head()

## 2.6. Transposing

# Transposing departments.csv:

df_dep.T

df_dep_t = df_dep.T

df_dep_t

# Resetting index in departments.csv:

df_dep_t.reset_index()

# Taking the first row of df_dep_t for the header:

new_header = df_dep_t.iloc[0]

new_header

# Deleting the first row (by creating a new df without it):

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Inserting new headers in df_dep_t_new:

df_dep_t_new.columns = new_header

df_dep_t_new

# 3. Subsetting

## 3.1. Creating Data Dictionary

# Creating data dictionary:

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

## 3.2. Creating Subset

# Using data dictionary to determine the department:

print(data_dict.get('19'))

# Creating subset for snacks department only:

df_prods['department_id'] == 19

df_prods[df_prods['department_id']==19]

df_snacks = df_prods[df_prods['department_id']==19]

df_snacks.head()

# Option 2 for snacks subset:

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

# Option 3 for snacks subset:

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3

# 4. Data Wrangling & Subsetting (Task)

## 4.1. Changing column data type and name

# Item 2. Changing data type of 'user_id' in orders.csv:

df_ords['user_id'] = df_ords['user_id'].astype('str')

type('user_id')

# Item 3. Changing the name of 'order_hour_of_day' column to 'time_of_order, hour':

df_ords.rename(columns = {'order_hour_of_day' : 'time_of_order, hour'}, inplace = True)

df_ords.head()

## 4.2. Finding busiest hour and Determining meaning behind a value

# Item 4. Finding busiest hour for placing orders - 10:00am:

df_ords['time_of_order, hour'].value_counts()

# Item 5. Determining the meaning behind a value of 4 in the "department_id" column within the df_prods:

print(data_dict.get('4'))

## 4.3. Creating subset and finding all observations

# Item 6. Creating a subset for breakfast item sales:

df_breakfast = df_prods.loc[df_prods['department_id'] == 14]

df_breakfast

# Item 7. Finding all observations for alcohol, deli, beverages, and meat/seafood departments:

df_alc_deli_bev_meat = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_alc_deli_bev_meat

## 4.4. Counting rows and and viewing all info + basic stats

# Item 8. Counting the rows in df_alc_deli_meat - 7650!:

df_alc_deli_bev_meat.info()

# Item 9. All info about "user_id: of "1":

df_user_id_1 = df_ords[df_ords['user_id'] == '1']

df_user_id_1

# Item 10. Basic stats on "user_id" "1":

df_user_id_1.describe()

# 5. Exporting 'orders_wrangled' and 'departments_wrangled'

# Item 12. Exporting df_ords as orders_wrangled.csv:

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# Item 13. Exporting df_dep_t_new as departments_wrangled.csv:

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data','departments_wrangled.csv'))
