==> Instacart Basket Analysis/02 Scripts/3. Data Consistency Checks.ipynb <==
# DATA CONSITENCY CHECKS

# 1. Importing pandas, NumPy, and os libraries

# 2. Importing 'products.csv' and 'orders_wrangled.csv'

# 3. Creating a dataframe with mixed type column and replacing it with string

# 4. Finding missing values in 'df_prods'

# 5. Finding mean and median values in 'df_prods'

# 6. Removing missing values in 'df_prods'

# 7. Finding and dropping duplicates in 'df_prods_clean'

# 8. Exporting 'products_checked' without duplicates

# 9. Data Consistency Checks (Task)

## 9.1. Describing 'df_prods'

## 9.2. Checking for mixedtype data in 'df_ords'

## 9.3. Addressing missing values in 'df_ords'

## 9.4. Checking for duplicates in 'df_ords'

## 9.5. Exporting 'df_ords' as 'orders_checked'

# ----------------------------------------------

# 1. Importing pandas, NumPy, and os libraries

# Importing the pandas, NumPy, and os libraries:

import pandas as pd
import numpy as np
import os

# 2. Importing 'products.csv' and 'orders_wrangled.csv'

# Defining path:

path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

# Importing products.csv:

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'))

# Importing orderswrangled.csv:

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'))

# 3. Creating a dataframe with mixed type column and replacing it with string

# Creating a dataframe:

df_test = pd.DataFrame()

# Creating a mixed type column:

df_test['mix'] = ['a', 'b', 1, True]
df_test.head()

# Checking for mixed types in df_test:

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Replacing mixed type with string type in df_test:

df_test['mix'] = df_test['mix'].astype('str')

# 4. Finding missing values in 'df_prods'

# Finding missing values in df_prods:

df_prods.isnull().sum()

# Creating a subset with missing values only:

df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

# 5. Finding mean and median values in 'df_prods'

# Finding the mean values for the columns with missing values:

df_prods.describe()

# Finding the median value in df_prods:

df_prods.median()

# Finding the number of rows and columns in df_prods:

df_prods.shape

# 6. Removing missing values in 'df_prods'

# Creating a new df_prods_clean without missing values:

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
df_prods_clean

df_prods_clean.shape

# 7. Finding and dropping duplicates in 'df_prods_clean'

# Finding duplicates in df_prods_clean:

df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

# Dropping duplicates in df_prods_clean:

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups

df_prods_clean_no_dups.shape

# 8. Exporting 'products_checked' without duplicates

# Exporting df_prods_clean_no_dups:

df_prods_clean_no_dups.to_csv(os.path.join(path,'02 Data','Prepared Data','products_checked.csv'))

# 9. Data Consistency Checks (Task)

## 9.1. Describing 'df_prods'

# Item 2. Running df.describe() on df_prods:

df_prods.describe()

    ## Interpreting the output of "df_prods.describe()":
    
df_prods.median()

#### 1. The max value for "product_id" (49688) is 5 points lower than the count (49693) meaning that there might be duplicates or missing values in the column
2. The max value for the "prices" being 99999.000000, while the median is only 7.1, gives us an idea that this 99999.00000 might be an outlier and needs to be dealt with accordingly.
3. The min value for "prices" is exactly 1.000000, which is suspicious, taking into account that there are usually multiple items in stores that are priced under $1.

## 9.2. Checking for mixedtype data in 'df_ords'

# Item 3. Checking for mixed-type data in 'df_ords':

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#### - No mixed values have been detected in df_ords.

## 9.3. Addressing missing values in 'df_ords'

# Item 5. Running a check for missing values in df_ords:

df_ords.shape

df_ords.isnull().sum()

#### - Findings and explanations for missing values: there are 206,209 missing values for "days_since_prior_order"

# Item 6. Addressing the missing values:

df_ords.describe().apply(lambda s: s.apply('{0:.5f}'.format))

    ## Calculating the median:
    
df_ords.median()

    ## Imputing the median instead of missing values:
    
df_ords['days_since_prior_order'].fillna(7, inplace=True)

    ## Checking if missing values are gone:
    
df_ords.isnull().sum()

    ## Checking the change in the mean for "days_since_prior_order":
    
df_ords.describe().apply(lambda s: s.apply('{0:.5f}'.format))

#### Why I used imputing median values method: 1. The number of missing values (206,209) was a considerable amount - around 6.02% of the total number of values in the column, so it could not be disregarded. 2. Imputing seemed like a pretty natural thing to do to deal with missing values for number of days between orders, and for missing data comprising 6% of the column data. 3. When choosing between imputing mean or median values, I chose median, since I did not want to lower the mean of the column too much.

## 9.4. Checking for duplicates in 'df_ords'

# Item 7. Checking for duplicate values in df_ords:

df_ords_dups = df_ords[df_ords.duplicated()]
df_ords_dups

#### - Findings and an explanation for duplicate values: There are no duplicates found in df_ords.

# Item 8. Addressing duplicate values in df_ords: There are no duplicates found in df_ords

#### - Explanation of the menthod: N/A

## 9.5. Exporting 'df_ords' as 'orders_checked'

# Item 9. Exporting df_ords:

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data','orders_checked.csv'))

df_ords.shape
