==> Instacart Basket Analysis/02 Scripts/4.1. Combining & Exporting Data 1.ipynb <==
# COMBINING & EXPORTING DATA, PART 1

# 1. Importing libraries

# 2. Importing 'order_products_prior' and 'orders_checked

# 3. Merging 'df_ords_prior' and 'df_ords' into 'df_merged_large'

# 4. Exporting 'df_merged_large' as 'orders_products_combined.pkl'

# -------------------------------------------

# 1. Importing libraries

# Import libraries:

import pandas as pd
import numpy as np
import os

# 2. Importing 'order_products_prior' and 'orders_checked

# Define path:

path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

# Import order_products_prior.csv:

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'order_products__prior.csv'))

df_ords_prior.shape

# Import orders_checked.csv:

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_checked.csv'))

# Check the output of df_ords_prior and df_ords:

df_ords_prior.head()

df_ords.head()

df_ords_prior.shape

df_ords.shape

# 3. Merging 'df_ords_prior' and 'df_ords' into 'df_merged_large'

# Item 1. Merge df_ords_prior with df_ords:

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

    ## Confirm the results of the merge using merge flag:
    
    df_merged_large['_merge'].value_counts()

    ## Check the columns of df_merged_large:
    
    df_merged_large.head()

# 4. Exporting 'df_merged_large' as 'orders_products_combined.pkl'

# Item 2. Export df_merged_large as orders_products_combined.pkl:

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_combined.pkl'))

df_merged_large.shape
