==> Instacart Basket Analysis/02 Scripts/4.2. Combining & Exporting Data 2.ipynb <==
# COMBINING & EXPORTING DATA, PART 2

# 1. Importing libraries

# 2. Importing 'orders_products_combined' and 'products_checked'

# 3. Merging 'df_orders_products_combined' and 'df_products' on 'product_id'

# 4. Exporting 'df_orders_products_merged'

# ---------------------------------------------------------

# 1. Importing libraries

# Import libraries:

import pandas as pd
import numpy as np
import os

# 2. Importing 'orders_products_combined' and 'products_checked'

#Define the path:

path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

# Import orders_products_combined:

df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_combined.pkl'))

# Check the shape of orders_product_combined:

df_orders_products_combined.shape

    ## Check the columns of df_orders_products_combined:
    
    df_orders_products_combined.head()

# Item 3. Import products_checked:

df_products = pd.read_csv(os.path.join(path, '02 Data','Prepared Data','products_checked.csv'))

# Item 4. Check the shape of df_products:
    
df_products.shape

    ## Check the columns of df_products:
    
    df_products.head()

# 3. Merging 'df_orders_products_combined' and 'df_products' on 'product_id'

# Item 5. Merge df_orders_products_combined and df_products on 'product_id':

df_orders_products_merged = df_orders_products_combined.merge(df_products, on = 'product_id')

# Item 6. Confirm the results of the merge using merge flag:
    
df_orders_products_merged['_merge'].value_counts()

    ## Check the shape of df_orders_products_merged:
    
    df_orders_products_merged.shape

# 4. Exporting 'df_orders_products_merged'

# Item 7. Export df_orders_products_merged:

df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_merged.pkl'))
