==> Instacart Basket Analysis/02 Scripts/5. Deriving New Variables.ipynb <==
# DERIVING NEW VARIABLES

# 1. Importing libraries

# 2. Importing 'orders_products_merged'

# 3. Sorting products into price ranges

  ## 3.1. Sorting a subset

### 3.1.1. ... using user defined function

### 3.1.2. ...using df.loc( )

## 3.2. Sorting entire 'ords_prods_merged' data frame using df.loc( )

# 4. Creating 'busiest_day' column

# 5. Creating 'busy_slow_days' column (2 busiest and 2 slowest days)

# 6. Creating 'busiest_period_of_day' column (with periods of time: 'Most orders', 'Average orders', and 'Fewest orders')

# 7. Export 'ords_prods_merged' as 'orders_products_merged_new_columns'

# ----------------------------------------------

# 1. Importing libraries

# Import libraries:

import pandas as pd
import numpy as np
import os

# 2. Importing 'orders_products_merged'

 # Define the path:
    
path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

# Import orders_products_merged:

ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_merged.pkl'))

# Check the shape of 'ords_prods_merged':

ords_prods_merged.shape

# Check the columns of 'ords_prods_merged':

ords_prods_merged.head()

# 3. Sorting products into price ranges

   ## 3.1. Sorting a subset

# Create 1 mil row subset of:

df = ords_prods_merged[:1000000]

# Check the shape of df:

df.shape

### 3.1.1. ... using user defined function

# Define a fuction:

def price_label(row):
    
    if row['prices'] <= 5:
        return 'Low range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid rage product'
    elif row['prices'] > 15:
        return 'High range product'
    else: return np.nan

# Apply the function in a new column 'price_range':

df['price_range'] = df.apply(price_label, axis = 1)

    # Check for results of the 'price_rage' flag:

    df['price_range'].value_counts(dropna = False)

    # Check for the highest price in the subset:

    df['prices'].max()

### 3.1.2. ...using df.loc( )

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low range product'

    # Count up the number of products within each label in the “price_range_loc” column, missing values included:

    df['price_range_loc'].value_counts(dropna = False)

## 3.2. Sorting entire 'ords_prods_merged' data frame using df.loc( )

ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High range product'

ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid range product'

ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low range product'

    # Count up the number of products within each label in the “price_range_loc” column, missing values included:

    ords_prods_merged['price_range_loc'].value_counts(dropna = False)

# 4. Creating 'busiest_day' column

# Find out on which day most orders take place:

ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Create a list that will contain values: “Busiest day,” “Least busy,” and “Regularly busy”:

result = []

for value in ords_prods_merged['orders_day_of_week']:
    
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regularly busy')

result

# Create a new column 'busiest_day' and populate it with 'result' values:

ords_prods_merged['busiest_day'] = result

ords_prods_merged['busiest_day'].value_counts(dropna = False)

# 5. Creating 'busy_slow_days' column (2 busiest and 2 slowest days)

# Find out how many orders take place on which day:

ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

#### - Days 0 and 1 are the two busiest days, and days 3 and 4 are the two slowest ones

# Create a list that contains values: 'Busiest days', 'Slowest days' and 'Regularly busy':

busyslow = []

for x in ords_prods_merged['orders_day_of_week']:
    
    if x == 0:
        busyslow.append('Busiest days')
    elif x == 1:
        busyslow.append('Busiest days')
    elif x == 3:
        busyslow.append('Slowest days')
    elif x == 4:
        busyslow.append('Slowest days')
    else:
        busyslow.append('Regularly busy')

busyslow

# Create a new column 'busy_slow_days' and populate it with 'busyslow' list:

ords_prods_merged['busy_slow_days'] = busyslow

# Item 3. Check values of 'busy_slow_days'

ords_prods_merged['busy_slow_days'].value_counts(dropna = False)

ords_prods_merged.head()

# 6. Creating 'busiest_period_of_day' column (with periods of time: 'Most orders', 'Average orders', and 'Fewest orders')

# Find out the busiest, quietest and average hours:

ords_prods_merged['time_of_order, hour'].value_counts(dropna = False)

#### - Hours 9 through 16 are the busiest, hours 23 through 6 are the quietest, the rest are average.

# Create a list that contains values: 'Most orders', 'Average orders', and 'Fewest orders'

busy_hrs = []

for y in ords_prods_merged['time_of_order, hour']:
    
    if y == 9 or y == 10 or y == 11 or y == 12 or y == 13 or y == 14 or y == 15 or y ==16:
        busy_hrs.append('Most orders')
    elif y == 23 or y == 0 or y == 1 or y == 2 or y == 3 or y == 4 or y == 5 or y == 6:
        busy_hrs.append('Fewest orders')
    else:
        busy_hrs.append('Average orders')

# Create a new column 'busiest_period_of_day' and populate it with 'busyslow' list:

ords_prods_merged['busiest_period_of_day'] = busy_hrs

# Item 5. Check values of 'busiest_period_of_day'

ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

# 7. Export 'ords_prods_merged' as 'orders_products_merged_new_columns'

# Item 6. Export ords_prods_merged:

ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_merged_new_columns.pkl'))
