==> Instacart Basket Analysis/02 Scripts/7.1. Data Visualization Part 1.ipynb <==
# INTRO TO DATA VISUALIZATION WITH PYTHON (Task) - 1

# 1. Installing libraries

# 2. Importing 'customers' dataframe as 'cust':

# 3. Wrangling of 'cust'

## 3.1. Removing 'date_joined' column

## 3.2. Renaming 'Surnam', 'First Name', 'n_dependants'

# 4. Consitency checks of 'cust' (missing values, duplicates, mixed types)

# 5. Combining 'customer' df with 'orders_products_merged_new_columns_and_flags.pkl' on 'user_id'

# 6. Exporting 'orders_products_cust_merged' as 'orders_products_customers_merged.pkl'

# ----------------------------------------------------

# 1. Installing libraries

# Item 3. Install libraries:

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 2. Importing 'customers' dataframe as 'cust':

    # Define the path
    
path = r'/Users/elena/Documents/Data Analytics Course/4. Python fundamentals for Data Analysts/Instacart Basket Analysis'

    # Import customers.csv:

cust = pd.read_csv(os.path.join(path, '02 Data','Original Data','customers.csv'))

cust.shape

cust.head()

# 3. Wrangling of 'cust'

## 3.1. Removing 'date_joined' column

    # Remove column 'date_joined' from 'cust':

cust = cust.drop(columns = ['date_joined'])

cust.head()

## 3.2. Renaming 'Surnam', 'First Name', 'n_dependants'

    # Rename 'Surnam' column into 'last_name':
    
cust.rename(columns = {'Surnam' : 'last_name'}, inplace = True)

    # Rename 'First Name' column into 'first_name':
    
cust.rename(columns = {'First Name' : 'first_name'}, inplace = True)

    # Rename 'n_dependants' column into 'no_of_dependants':
    
cust.rename(columns = {'n_dependants' : 'no_of_dependants'}, inplace = True)

cust.head()

# 4. Consitency checks of 'cust' (missing values, duplicates, mixed types)

# Check for missing values in 'cust':
    
cust.isnull().sum()

# Check for duplicates in 'cust':
    
cust_dups = cust[cust.duplicated()]
cust_dups

##### - No duplicates founds in 'cust'.

# Check for mixed types in'cust':
    
for col in cust.columns.tolist():
  weird = (cust[[col]].applymap(type) != cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (cust[weird]) > 0:
        print (col)

type('first_name')

    # Replace mixed type in 'first_name' column in 'cust' df with string type:
    
cust['first_name'] = cust['first_name'].astype('str')

# 5. Combining 'customer' df with 'orders_products_merged_new_columns_and_flags.pkl' on 'user_id'

# Convert 'user_id' data type in 'cust' to string:
    
cust['user_id'] = cust['user_id'].astype('str')

cust['user_id'].dtype

cust.shape

# Import 'orders_products_merged_new_columns_and_flags':
    
orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_merged_new_columns_and_flags.pkl'))

# Check the shape of 'orders_products_merged':
    
orders_products_merged.shape

# Check the data type of 'user_id' in 'orders_products_merged':
    
orders_products_merged['user_id'].dtype

# Convert 'user_id' data type in 'orders_products_merged' to string:
    
orders_products_merged['user_id'] = orders_products_merged['user_id'].astype('str')

# Drop '_merge' in 'orders_products_merged':
    
orders_products_merged = orders_products_merged.drop(columns = ['_merge'])

# Merge 'cust' with 'orders_products_merged_new_columns_and_flags':

orders_products_cust_merged = cust.merge(orders_products_merged, on = 'user_id', indicator = True)

orders_products_cust_merged.head()

# Check the shape of the new 'orders_products_cust_merged':

orders_products_cust_merged.shape

orders_products_cust_merged.columns

# Confirm the results of the merge using merge flag:
    
orders_products_cust_merged['_merge'].value_counts()

# 6. Exporting 'orders_products_cust_merged' as 'orders_products_customers_merged.pkl'

# Item 8. Export 'orders_products_cust_merged':

orders_products_cust_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data','orders_products_customers_merged.pkl'))
